Cross-matching function + int64 types

+++ yael/hamming.c
@@ -160,9 +160,9 @@ hammatch_t * hammatch_realloc (hammatch_t * m, int n)
 
 
 void match_hamming_thres (const uint8 * qbs, const uint8 * dbs, int nb, int ht,
-                          long bufsize, hammatch_t ** hmptr, long * nptr)
+                          int64 bufsize, hammatch_t ** hmptr, int64 * nptr)
 {
-  long j, posm = 0;
+  int64 j, posm = 0;
   uint16 h;
   *hmptr = hammatch_new (bufsize);
   hammatch_t * hm = *hmptr;
@@ -194,10 +194,11 @@ void match_hamming_thres (const uint8 * qbs, const uint8 * dbs, int nb, int ht,
 }
 
 
-void match_hamming_thres_generic (const uint8 * qbs, const uint8 * dbs, int nb, int ht,
-                                  long bufsize, hammatch_t ** hmptr, long * nptr, int ncodes)
+void match_hamming_thres_generic (const uint8 * qbs, const uint8 * dbs, 
+                                  int nb, int ht, int64 bufsize, 
+                                  hammatch_t ** hmptr, int64 * nptr, int64 ncodes)
 {
-  int j, posm = 0;
+  int64 j, posm = 0;
   uint16 h;
   *hmptr = hammatch_new (bufsize);
   hammatch_t * hm = *hmptr;
@@ -230,9 +231,9 @@ void match_hamming_thres_generic (const uint8 * qbs, const uint8 * dbs, int nb, 
 
 
 void crossmatch_he (const uint8 * dbs, long n, int ht,
-                    long bufsize, hammatch_t ** hmptr, long long * nptr)
+                    long bufsize, hammatch_t ** hmptr, int64 * nptr)
 {
-  long long i, j, posm = 0;
+  int64 i, j, posm = 0;
   uint16 h;
   *hmptr = hammatch_new (bufsize);
   hammatch_t * hm = *hmptr;
@@ -272,9 +273,9 @@ void crossmatch_he (const uint8 * dbs, long n, int ht,
 
 
 
-void crossmatch_he_count (const uint8 * dbs, int n, int ht, long long * nptr)
+void crossmatch_he_count (const uint8 * dbs, int n, int ht, int64 * nptr)
 {
-  long long i, j, posm = 0;
+  int64 i, j, posm = 0;
   const uint8 * bs1 = dbs;
   
   for (i = 0 ; i < n ; i++) {
@@ -300,9 +301,10 @@ void crossmatch_he_count (const uint8 * dbs, int n, int ht, long long * nptr)
 /*-------------------------------------------*/
 /* Threaded versions, if OpenMP is available */
 #ifdef _OPENMP
-void compute_hamming_thread (uint16 * dis, const uint8 * a, const uint8 * b, int na, int nb)
+void compute_hamming_thread (uint16 * dis, const uint8 * a, const uint8 * b, 
+                             int na, int nb)
 {
-  int i, j;
+  int64 i, j;
 #pragma omp parallel shared (dis, a, b, na, nb) private (i, j)
     {
 #pragma omp for
